@@ -13,6 +13,6 @@ namespace HexaGen.CppAst.Model.Interfaces
         /// <summary>
         /// Gets or sets the comment attached to this element. Might be null.
         /// </summary>
-        CppComment Comment { get; set; }
+        CppComment? Comment { get; set; }
     }
 }
